requirements.txt
asyncio==3.4.3
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
DateTime==4.5
distlib==0.3.4
filelock==3.7.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Ext==0.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
platformdirs==2.5.2
pycparser==2.21
pytz==2022.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
SQLAlchemy==1.4.39
virtualenv==20.15.1
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
zope.interface==5.4.0
